Not tested- - adjusted encoder calibration so i can toggle based on commutation type - Made direction a preprocessor directive - made deadtime > 50ns Depending on Htimer resolution - added toggle between 12B and 6B commutations

+++ code/x694_BlockComm/x694_espidf_6Block_MCPWM/src/GateControl.h
@@ -82,15 +82,15 @@ inline mcpwm_generator_config_t HPWMSetup = {
 
 #define DT_SCALER (HighTimerResolution * 1.0 / 1e9)
 const mcpwm_dead_time_config_t highGateDeadTimeSetup = {
-    .posedge_delay_ticks = ceil(isrTickDeadTime * DT_SCALER),
-    .negedge_delay_ticks = ceil(isrTickDeadTime * DT_SCALER),
+    .posedge_delay_ticks = (uint32_t) ceil(isrTickDeadTime * DT_SCALER),
+    .negedge_delay_ticks = (uint32_t) ceil(isrTickDeadTime * DT_SCALER),
     .flags = {
         // invert_output = 1;
     }
 };
 const mcpwm_dead_time_config_t lowGateDeadTimeSetup = {
-    .posedge_delay_ticks = ceil( (isrTickDeadTime + relativeDeadTime) * DT_SCALER),
-    .negedge_delay_ticks = ceil( isrTickDeadTime * DT_SCALER),
+    .posedge_delay_ticks = (uint32_t) ceil( (isrTickDeadTime + relativeDeadTime) * DT_SCALER),
+    .negedge_delay_ticks = (uint32_t) ceil( isrTickDeadTime * DT_SCALER),
     .flags = {
         // invert_output = 1;
     }
